instatiated 3 new person objects and tested the creation of one with a constructor that didn't exsit

+++ Java e POO - 2/exercicio-pratico-1/src/com/company/Pessoa.java
@@ -6,8 +6,8 @@ public class Pessoa {
     private String nome;
     private Integer idade;
     private String id;
-    private float peso;
-    private float altura;
+    private double peso;
+    private double altura;
 
 
 // Exercício 2
@@ -22,7 +22,7 @@ public class Pessoa {
         this.id = id;
     }
 
-    public Pessoa(String nome, Integer idade, String id, float peso, float altura) {
+    public Pessoa(String nome, Integer idade, String id, double peso, double altura) {
         this.nome = nome;
         this.idade = idade;
         this.id = id;
